Moved testing functionality into a helper class.
