+++ Source/Application.js
@@ -14,6 +14,7 @@ import { Image } from './Image.js'
 import { Shape } from './Shape.js'
 import { Artboard } from './Artboard.js'
 import { Page } from './Page.js'
+import { Tester } from './Tester.js'
 
 /**
     Gives you access to Sketch, and provides access to:
@@ -326,41 +327,8 @@ export class Application extends WrappedObject {
 
     runUnitTests() {
       var tests = Application.tests(this);
-      var ran = 0;
-      var passes = [];
-      var failures = [];
-      var testFailures;
-      var tester = {
-        "assert" : function(value, description) {
-          if (!value) {
-            if (!description) description = "";
-            testFailures.push(description)
-          }
-        },
-        "assertEqual" : function(v1, v2) {
-          if (v1 != v2) {
-            testFailures.push(v1 + " != " + v2)
-          }
-        }
-      }
-      var test;
-      for (test in tests) {
-        ran++;
-        testFailures = [];
-        var result = tests[test](tester);
-        if (testFailures.length > 0) {
-          failures.push({"name" : test, "reasons" : testFailures})
-        } else {
-          passes.push(test)
-        }
-      }
-
-      return {
-        "ran" : ran,
-        "crashed" : 0,
-        "failures" : failures,
-        "passes" : passes
-      };
+      var tester = new Tester();
+      return tester.runUnitTests(tests);
     }
 
 }
